Purple- movieDetail.jsp 별점버그수정, reviewList.jsp 버그수정 (#113)
* home.jsp 최신영화 onclick 수정

* 카카오 로그인 구현중

* kakao login 구현중- 로그인성공

* LoginContoller.java, movieDetail.jsp 수정

* member-mapping.xml  수정, kakao 소셜작업중

* kakao login 회원가입 완성

* kakao 연동 완성

* 좋아요 기능 시작

* 좋아요 기능 작업중

* updateLikeCount dao, service, method 추가

* 좋아요 기능 작업중

* 좋아요 기능 작업중

* 좋아요insert, delete, update 기능 완성, 숫자 ajax 구현중

* reviewList.jsp css 수정

* reviewList.jsp css 추가 수정

* reviewList.jsp에서 상세페이지로 이동 기능 추가

* reviewList.jsp에 로그인한 멤버의 리뷰만 보이도록 수정

* 좋아요 기능 완성

* 좋아요 기능 완성

* review-mapping.xml 수정

* review-mapping.xml 재수정

* paymentList.jsp 기능 작업중

* paymentList.jsp css, 데이터불러오기 완성

* movieDetail.jsp 별점버그수정, reviewList.jsp 버그수정

+++ src/main/java/com/main/oneflix/review/client/controller/ReviewController.java
@@ -30,15 +30,12 @@ public class ReviewController {
 	private DeleteReviewService deleteReviewService;
 	@Autowired
 	private GetReviewListService getReviewListService;
-	@Autowired
-	private GetReviewLikeListService getReviewLikeService;
 
 	@RequestMapping("/insertReviewProc.do")
 	public ModelAndView insertReviewProc(ReviewVO vo, HttpSession session, ModelAndView mav) {
 		MemberVO member = (MemberVO) session.getAttribute("member");
 		vo.setEmail(member.getEmail());
-		if (vo.getReviewId() != null)
-			deleteReviewService.deleteReview(vo);
+		if (vo.getReviewId() != null) deleteReviewService.deleteReview(vo);
 		insertReviewService.insertReview(vo);
 		mav.addObject("movieId", vo.getMovieId());
 		mav.setViewName("redirect:/getMovieDetailProc.do");
